refactore frontend and backend, add Readme but still just backend, add seeding data

+++ server/src/controllers/productType.controller.ts
@@ -1,6 +1,9 @@
 import { Request, Response } from "express";
 import productTypeService from "../services/productType.service.js";
 import { StatusCodes } from "../utils/constant.js";
+import { IRequestExtends } from "../utils/express-extends.js";
+import { checkRole } from "../utils/check-role.js";
+import { UnauthenticatedError } from "../errors/index.error.js";
 
 const getAndMap = async (): Promise<unknown> => {
     const productTypes = await productTypeService.getAllPopulateCategory();
@@ -31,7 +34,13 @@ const getByName = async (
     return;
 };
 
-const add = async (req: Request, res: Response): Promise<void> => {
+const add = async (req: IRequestExtends, res: Response): Promise<void> => {
+    const user = req.user!;
+    if (checkRole(user.role, "user")) {
+        throw new UnauthenticatedError(
+            "User does not have the right permission"
+        );
+    }
     await productTypeService.addAsync(req.body);
 
     res.status(StatusCodes.Created201).send({
@@ -40,15 +49,28 @@ const add = async (req: Request, res: Response): Promise<void> => {
     return;
 };
 
-const remove = async (req: Request, res: Response): Promise<void> => {
+const remove = async (req: IRequestExtends, res: Response): Promise<void> => {
+    const user = req.user!;
+    if (!checkRole(user.role, "admin")) {
+        throw new UnauthenticatedError(
+            "User does not have the right permission"
+        );
+    }
+
     await productTypeService.deleteAsync(req.params.id);
 
     res.status(StatusCodes.Ok200).send({ productTypes: await getAndMap() });
     return;
 };
 
-const update = async (req: Request, res: Response): Promise<void> => {
-    console.log(req.body);
+const update = async (req: IRequestExtends, res: Response): Promise<void> => {
+    // console.log(req.body);
+    const user = req.user!;
+    if (checkRole(user.role, "user")) {
+        throw new UnauthenticatedError(
+            "User does not have the right permission"
+        );
+    }
     await productTypeService.updateAsync(req.params.id, req.body);
 
     res.status(StatusCodes.Ok200).send({ productTypes: await getAndMap() });
